+++ src/Shodan.FSharp/AssemblyInfo.fs
@@ -6,8 +6,6 @@ open System.Runtime.InteropServices
 [<assembly: AssemblyTitleAttribute("Shodan.FSharp")>]
 [<assembly: AssemblyProductAttribute("Shodan.FSharp")>]
 [<assembly: AssemblyDescriptionAttribute("A Shodan API client for F#.")>]
-[<assembly: AssemblyVersionAttribute("0.1.4")>]
-[<assembly: AssemblyFileVersionAttribute("0.1.4")>]
 [<assembly: GuidAttribute("cc90107d-7c65-4dcb-bc13-bb3e0d75933f")>]
 do ()
 
@@ -15,6 +13,4 @@ module internal AssemblyVersionInformation =
     let [<Literal>] AssemblyTitle = "Shodan.FSharp"
     let [<Literal>] AssemblyProduct = "Shodan.FSharp"
     let [<Literal>] AssemblyDescription = "A Shodan API client for F#."
-    let [<Literal>] AssemblyVersion = "0.1.4"
-    let [<Literal>] AssemblyFileVersion = "0.1.4"
     let [<Literal>] Guid = "cc90107d-7c65-4dcb-bc13-bb3e0d75933f"