+++ RELEASE_NOTES.md
@@ -29,8 +29,19 @@
 * Fix: URI concatenation works
 * Added: more type modeling to web APIs
 
-## 0.1.4 - June 4 2017
+### 0.1.4 - June 4 2017
 
 * Added typed queries for telnet options
 * Changed the result type of DNS queries
 * Forward non-200 HTTP statuses through 'ShodanException'
+
+### 0.2.0 - June 5 2017
+
+* Change the web API types to instance classes taking a `SecureString` API key
+* Module `WebApi` changed to `Http`
+* `ShodanError` changed to `ShodanWebException`
+* Remove ILMerge
+
+### 0.3.0 - July 18 2017
+* Changed the banner type returned by search queries
+* Use .NET 4.5